Allow customization of the initial solution.
By default, if TrajOptProblem is given an initial state, the initial
solution is initialized with this initial state for the first element,
and with the neutral state for the others. Both these states can have
very different values, yielding poor starting values for xs, prev_xs
and trial_xs. In the classical case where all the states are in the
same stage space, one might want to start with a constant solution.
Instead of trying to detect if this is a desirable behavior, or adding
a configuration parameter, an initialization callback is added to the
problem, which can be set by the user before calling the solver
setup(). If it is not set, the previous default behavior applies.